+++ Makefile.in
@@ -77,6 +77,7 @@ OBJZ = \
 	compare258.o \
 	compress.o \
 	crc32.o \
+	crc32_comb.o \
 	deflate.o \
 	deflate_fast.o \
 	deflate_medium.o \
@@ -106,6 +107,7 @@ PIC_OBJZ = \
 	compare258.lo \
 	compress.lo \
 	crc32.lo \
+	crc32_comb.lo \
 	deflate.lo \
 	deflate_fast.lo \
 	deflate_medium.lo \
